Move condition impl to plan files
