--- supplementary/alica_ros1/alica_ros_turtlesim/alica_turtlesim_library/include/conditions.h
@@ -1,17 +1,0 @@
-#pragma once
-#include <boost/dll/alias.hpp>
-
-namespace alica
-{
-class Blackboard;
-class RunningPlan;
-
-bool conditionMove2Init(const Blackboard* input, const RunningPlan* rp, const Blackboard* globalBlackboard);
-bool conditionInit2Move(const Blackboard* input, const RunningPlan* rp, const Blackboard* globalBlackboard);
-bool conditionDefaultCondition(const Blackboard* input, const RunningPlan* rp, const Blackboard* globalBlackboard);
-
-BOOST_DLL_ALIAS(alica::conditionMove2Init, Move2Init)
-BOOST_DLL_ALIAS(alica::conditionInit2Move, Init2Move)
-BOOST_DLL_ALIAS(alica::conditionDefaultCondition, DefaultCondition)
-
-} /* namespace alica */
