+++ supplementary/alica_ros1/alica_ros_turtlesim/alica_turtlesim_library/src/Master.cpp
@@ -1,5 +1,7 @@
-#include "Master.h"
 #include <engine/DefaultUtilityFunction.h>
+
+#include "Master.h"
+#include "world_model.hpp"
 
 namespace alica
 {
@@ -16,6 +18,18 @@ std::shared_ptr<UtilityFunction> MasterUtilityFunction::getUtilityFunction(Plan*
 {
     std::shared_ptr<UtilityFunction> defaultFunction = std::make_shared<DefaultUtilityFunction>(plan);
     return defaultFunction;
+}
+
+bool conditionMove2Init(const Blackboard* input, const RunningPlan* rp, const Blackboard* globalBlackboard)
+{
+    std::shared_ptr<turtlesim::ALICATurtleWorldModel> wm =
+            LockedBlackboardRO(*globalBlackboard).get<std::shared_ptr<turtlesim::ALICATurtleWorldModel>>("worldmodel");
+    return wm->getInit();
+}
+
+bool conditionInit2Move(const Blackboard* input, const RunningPlan* rp, const Blackboard* globalBlackboard)
+{
+    return rp->isAnyChildStatus(PlanStatus::Success);
 }
 
 } // namespace alica
